fix: serialize diagnostic shapes as int64 to avoid int32 overflow
Full-particle output wrote a negative shape when the per-species count exceeded INT32_MAX (3,175,598,925 wrapped to -1119368371). Widen the nixio put_metadata/get_metadata dims to int64 and fix the particle and tracer count narrowing; add a regression test round-tripping >2^32 counts. Also document the repo's git hooks in AGENTS.md.

+++ pic/diag/field.hpp
@@ -154,7 +154,7 @@ public:
       const char name[]  = "uf";
       const char desc[]  = "electromagnetic field";
       int        ndim    = 5;
-      int        dims[5] = {nc, nz, ny, nx, 6};
+      int64      dims[5] = {nc, nz, ny, nx, 6};
       nixio::put_metadata(dataset, name, "f8", desc, disp0, nbyte, ndim, dims);
     }
 
@@ -174,7 +174,7 @@ public:
       const char name[]  = "um";
       const char desc[]  = "moment";
       int        ndim    = 6;
-      int        dims[6] = {nc, nz, ny, nx, Ns, 14};
+      int64      dims[6] = {nc, nz, ny, nx, Ns, 14};
       nixio::put_metadata(dataset, name, "f8", desc, disp0, nbyte, ndim, dims);
     }
