+++ pic/diag/tracer.hpp
@@ -68,10 +68,10 @@ public:
         std::string name = fmt::format("up{:02d}", species);
         std::string desc = fmt::format("tracer particle species {:02d}", species);
 
-        const int size    = ParticleType::get_particle_size();
-        const int Np      = nbyte / size;
-        const int ndim    = 2;
-        const int dims[2] = {Np, ParticleType::Nc};
+        const int   size    = ParticleType::get_particle_size();
+        const int64 Np      = nbyte / size;
+        const int   ndim    = 2;
+        const int64 dims[2] = {Np, ParticleType::Nc};
 
         nixio::put_metadata(dataset, name, "f8", desc, disp0, nbyte, ndim, dims);
       }
